Feature/task 4

+++ TodoApp/src/TodoApp.Api/Controllers/ItemsController.cs
@@ -15,22 +15,26 @@ namespace TodoApp.Api.Controllers
         private const string InvalidText = "Invalid item text";
         private const string NonEmptyId = "Id of item can't be set";
         private const string EmptyId = "Id can't be empty";
+        private const string InvalidIds = "Id in url and body does not match";
 
         private readonly IAddItemService _addItemService;
         private readonly IGetItemByIdService _getItemByIdService;
+        private readonly IUpdateItemService _updateItemService;
         private readonly IItemRepository _repository;
         private readonly ILocationHelper _locationHelper;
 
         public ItemsController(
             IAddItemService addItemService,
             IGetItemByIdService getItemByIdService,
+            IUpdateItemService updateItemService,
             IItemRepository repository,
             ILocationHelper helper)
         {
             _addItemService = addItemService;
+            _getItemByIdService = getItemByIdService;
+            _updateItemService = updateItemService;
             _repository = repository;
             _locationHelper = helper;
-            _getItemByIdService = getItemByIdService;
         }
 
         public async Task<IHttpActionResult> GetAsync()
@@ -47,12 +51,12 @@ namespace TodoApp.Api.Controllers
 
             var retrievedItem = await _getItemByIdService.GetItemByIdAsync(id);
 
-            if (retrievedItem.WasFound)
+            if (!retrievedItem.WasFound)
             {
-                return Ok(retrievedItem.Entity);
+                return NotFound();
             }
 
-            return NotFound();
+            return Ok(retrievedItem.Entity);
         }
 
         public async Task<IHttpActionResult> PostAsync([FromBody] Item item)
@@ -71,20 +75,72 @@ namespace TodoApp.Api.Controllers
         }
 
         public async Task<IHttpActionResult> PutAsync(Guid id, [FromBody] Item item)
-            => Ok(await _repository.UpdateAsync(item));
+        {
+            ValidatePutArguments(id, item);
+
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var updatedItem = await _updateItemService.UpdateItemAsync(item);
+
+            if (!updatedItem.WasFound)
+            {
+                return NotFound();
+            }
+
+            return Ok(updatedItem.Entity);
+        }
 
         public async Task<IHttpActionResult> DeleteAsync(Guid id)
         {
+            ValidateGuid(id);
+
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var itemToDelete = await _getItemByIdService.GetItemByIdAsync(id);
+
+            if (!itemToDelete.WasFound)
+            {
+                return NotFound();
+            }
+
             await _repository.DeleteAsync(id);
 
             return StatusCode(HttpStatusCode.NoContent);
         }
 
-        private void ValidateGuid(Guid id)
+        private void ValidatePutArguments(Guid id, Item item)
         {
-            if (id == Guid.Empty)
+            if (item == null)
+            {
+                ModelState.AddModelError("NullItem", InvalidRequestBody);
+                return;
+            }
+
+            if (item.Id != id)
+            {
+                ModelState.AddModelError(nameof(Item.Id), InvalidIds);
+                return;
+            }
+
+            ValidatePutItem(item);
+        }
+
+        private void ValidatePutItem(Item item)
+        {
+            if (item.Id == Guid.Empty)
             {
                 ModelState.AddModelError(nameof(Item.Id), EmptyId);
+            }
+
+            if (!IsTextValid(item.Text))
+            {
+                ModelState.AddModelError(nameof(Item.Text), InvalidText);
             }
         }
 
@@ -96,10 +152,10 @@ namespace TodoApp.Api.Controllers
                 return;
             }
 
-            ValidateItem(item);
+            ValidatePostItem(item);
         }
 
-        private void ValidateItem(Item item)
+        private void ValidatePostItem(Item item)
         {
             if (item.Id != Guid.Empty)
             {
@@ -109,6 +165,14 @@ namespace TodoApp.Api.Controllers
             if (!IsTextValid(item.Text))
             {
                 ModelState.AddModelError(nameof(Item.Text), InvalidText);
+            }
+        }
+
+        private void ValidateGuid(Guid id)
+        {
+            if (id == Guid.Empty)
+            {
+                ModelState.AddModelError(nameof(Item.Id), EmptyId);
             }
         }
